Contains onboarding view controller with added code for location permission
