+++ Section 8/HelperFiles/locationController.swift
@@ -1,0 +1,33 @@
+//
+//  locationController.swift
+//  Section 8
+//
+//  Created by Ivan Ramirez on 12/10/18.
+//  Copyright © 2018 Warren. All rights reserved.
+//
+
+import Foundation
+import CoreLocation
+import UserNotifications
+
+
+class locationController {
+    
+//    let oneMile = Measurement(value: 1, unit: UnitLength.miles)
+    
+//    let radius = oneMile
+        //oneMile.converted(to: .meters).value
+//    let coordinate = CLLocationCoordinate2D(latitude: 37.33182,
+//                                            longitude: -122.03118)
+//    let region = CLCircularRegion(center: coordinate,
+//                                  radius: radius,
+//                                  identifier: UUID().uuidString)
+//
+//    region.notifyOnExit = false
+//    region.notifyOnEntry = true
+//
+//    let trigger = UNLocationNotificationTrigger(region: region,
+//        repeats: false)
+
+    
+}
