+++ Section 8/ViewControllers/stepDetailExtension.swift
@@ -16,9 +16,7 @@ import UserNotifications
 
 extension StepDetailTVC: DatePickerTVCellDelegate {
     
-    
-    
-    
+
     func datePickerButtonTapped(_ sender: DatePickerTVCell, _ picker: UIDatePicker) {
         print("\n\n🚀Set Time Button Tapped in: DatePickerTVCell\n")
